First swing at combatting Spatial Confounding in Count data using TPRS and Poisson Regression. The desired b1 and desired Point Esimate I believe to be b1 of a Poisson Regression without confounding (seen in the visualizer: red it what we are aiming for blue is the line with confounding).
Ceiling is used to get count data. In a very rare case < 1/1000, a negative number is gotten a special combination of residual error and the sample, so abs is used.

First run through took forever because calculating CI's on glm's takes long come to find out. You can see in PoissonRegB1est.pdf that I actually overshot the b1 of the desired model.